refactor(entity): Move entity classes into Entity namespace
